Split views/sessions.py into a package
The single-file view module had grown to 565 LOC bundling 3 serializers,
3 pydantic request models, and 7 view callables. Break it into focused
submodules so each lifecycle phase reads in isolation:

- views/sessions/__init__.py   re-exports the public surface
- views/sessions/serializers.py  resource/session/turn dict builders
- views/sessions/schemas.py    GitHubRepoResource / RunRequest / PromptRequest
- views/sessions/create.py     sessions_list_create + helpers
- views/sessions/lifecycle.py  get/prompt/turns/terminate/delete
- views/sessions/stream.py     SSE replay endpoint

No behavior change. urls.py imports unchanged. The pydantic models keep
__module__ pointing at agent_on_demand.views.sessions so the
request-schema discovery test and snapshot key continue to resolve.
Functions that tests patch at the package namespace
(check_can_accept_prompt, stream_session_from_db, session_service.*) are
re-exported there and the lifecycle/stream submodules dereference them
via the package at call time, so existing patches still take effect
without test changes.

scripts/scope_e2e.py: widen the views/sessions.py rule prefix to the
directory so changes to the new submodules still pull in test_sessions.py.
